requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
Werkzeug==2.2.3
whatthepatch==1.0.2
wheel==0.38.4